wsd: extract AI chat into its own AIChatSession class
ClientSession had ~1950 lines of AI chat code - the LLM tool loop,
tool dispatchers, image generation, approval gating, and kit-response
intercepts - mixed in with tile flow, clipboard, saveAs, etc. Pull it
out into a new class that ClientSession owns and delegates to for the
aichat:/aichatcancel:/aichatapprove: messages and five kit responses.

Pure move, no behavior changes.
