+++ wsd/ClientSession.cpp
@@ -37,7 +37,7 @@
 #include <wsd/DocumentBroker.hpp>
 #include <wsd/FileServer.hpp>
 #if !MOBILEAPP
-#include <wsd/DocumentToolDescriptions.hpp>
+#include <wsd/AIChatSession.hpp>
 #endif
 #include <wsd/TileDesc.hpp>
 
@@ -63,30 +63,6 @@ using namespace COOLProtocol;
 
 static constexpr float TILES_ON_FLY_MIN_UPPER_LIMIT = 10.0;
 static constexpr int SYNTHETIC_COOL_PID_OFFSET = 10000000;
-static constexpr std::size_t MAX_AI_IMAGE_GENERATIONS = 20;
-
-#if !MOBILEAPP
-static bool isValidImageSize(const std::string& size)
-{
-    const auto pos = size.find('x');
-    if (pos == std::string::npos || pos == 0 || pos == size.size() - 1)
-        return false;
-    const std::string widthStr = size.substr(0, pos);
-    const std::string heightStr = size.substr(pos + 1);
-    try
-    {
-        const int width = std::stoi(widthStr);
-        const int height = std::stoi(heightStr);
-        return width > 0 && height > 0
-            && widthStr == std::to_string(width)
-            && heightStr == std::to_string(height);
-    }
-    catch (const std::exception&)
-    {
-        return false;
-    }
-}
-#endif
 
 using Poco::Path;
 
@@ -162,6 +138,10 @@ ClientSession::ClientSession(const std::shared_ptr<ProtocolHandlerInterface>& ws
         std::to_string(ProcUtil::getProcessId() + SYNTHETIC_COOL_PID_OFFSET) + ",\"tid\":1},\n");
 
     _browserSettingsJSON = new Poco::JSON::Object();
+
+#if !MOBILEAPP
+    _aiChat = std::make_unique<AIChatSession>(*this);
+#endif
 }
 
 // Can't take a reference in the constructor.
@@ -595,1649 +575,6 @@ makeSignatureActionSession(std::shared_ptr<ClientSession> clientSession,
     return httpSession;
 }
 } // namespace
-
-void ClientSession::sendAIChatResult(bool success, const std::string& text,
-                                     const std::string& requestId)
-{
-    Poco::JSON::Object::Ptr result = new Poco::JSON::Object();
-    result->set("success", success);
-    if (success)
-        result->set("content", text);
-    else
-        result->set("error", text);
-    result->set("requestId", requestId);
-
-    std::ostringstream oss;
-    result->stringify(oss);
-    sendTextFrame("aichatresult: " + oss.str());
-}
-
-std::string ClientSession::mapAIHttpStatusToError(
-    http::StatusCode statusCode, const std::string& reasonPhrase,
-    const std::string& context)
-{
-    switch (statusCode)
-    {
-        case http::StatusCode::BadRequest:
-            return context.empty() ? "Invalid request"
-                                   : "Invalid " + context + " request";
-        case http::StatusCode::Unauthorized:     return "Invalid API key";
-        case http::StatusCode::Forbidden:        return "API key lacks permissions";
-        case http::StatusCode::TooManyRequests:  return "Rate limited - please wait a moment and retry";
-        case http::StatusCode::InternalServerError: return "API server error - try again later";
-        case http::StatusCode::ServiceUnavailable:  return "Service temporarily unavailable";
-        default:
-        {
-            std::string err = "API error (";
-            err.append(std::to_string(static_cast<int>(statusCode)));
-            err.append("): ");
-            err.append(reasonPhrase);
-            return err;
-        }
-    }
-}
-
-static const std::string AI_SYSTEM_PROMPT =
-    "You are a helpful assistant for Collabora Online. "
-    "Help users with their documents — answering questions, suggesting edits, "
-    "rewriting text, and more. When the user shares selected text from their document, "
-    "provide relevant help with that text. When no selected text is provided, answer "
-    "general questions about documents, formatting, writing, and the application. "
-    "When providing rewritten or edited text, return it in markdown format preserving "
-    "the original formatting structure. IMPORTANT: Return the markdown text directly "
-    "without wrapping it in code fences (do NOT use ```markdown or ``` blocks). "
-    "Just return the raw markdown content. Be concise and helpful.";
-
-namespace
-{
-
-/// Helper to create an OpenAI function-calling tool object.
-Poco::JSON::Object::Ptr makeAITool(const std::string& name,
-                                    const std::string& description,
-                                    Poco::JSON::Object::Ptr parameters)
-{
-    Poco::JSON::Object::Ptr fn = new Poco::JSON::Object();
-    fn->set("name", name);
-    fn->set("description", description);
-    fn->set("parameters", parameters);
-
-    Poco::JSON::Object::Ptr tool = new Poco::JSON::Object();
-    tool->set("type", "function");
-    tool->set("function", fn);
-    return tool;
-}
-
-/// Helper to create an OpenAI function parameter schema object.
-Poco::JSON::Object::Ptr makeParamSchema(
-    std::initializer_list<std::pair<std::string, std::pair<std::string, std::string>>> props,
-    std::initializer_list<std::string> required)
-{
-    Poco::JSON::Object::Ptr properties = new Poco::JSON::Object();
-    for (const auto& p : props)
-    {
-        Poco::JSON::Object::Ptr prop = new Poco::JSON::Object();
-        prop->set("type", p.second.first);
-        prop->set("description", p.second.second);
-        properties->set(p.first, prop);
-    }
-
-    Poco::JSON::Array::Ptr reqArr = new Poco::JSON::Array();
-    for (const auto& r : required)
-        reqArr->add(r);
-
-    Poco::JSON::Object::Ptr schema = new Poco::JSON::Object();
-    schema->set("type", "object");
-    schema->set("properties", properties);
-    schema->set("required", reqArr);
-    return schema;
-}
-
-} // anonymous namespace
-
-Poco::JSON::Array::Ptr ClientSession::buildAIToolDefinitions() const
-{
-    Poco::JSON::Array::Ptr tools = new Poco::JSON::Array();
-
-    // generate_image - existing tool
-    tools->add(makeAITool(
-        "generate_image",
-        "Generate an image based on the user's description. Call this when the "
-        "user asks to create, draw, generate, sketch, or make an image or picture.",
-        makeParamSchema(
-            {{"prompt", {"string", "A detailed description of the image to generate"}}},
-            {"prompt"})));
-
-    // extract_document_structure - inspect the open document
-    tools->add(makeAITool(
-        "extract_document_structure",
-        DocumentToolDescriptions::EXTRACT_DOC_STRUCTURE_DESCRIPTION,
-        makeParamSchema(
-            {{"filter", {"string",
-                "Filter results to a specific structure type. "
-                "For Impress: 'slides'. For Writer: 'contentcontrol'. "
-                "Omit to get the full structure."}}},
-            {})));
-
-    // transform_document_structure - modify the open document
-    tools->add(makeAITool(
-        "transform_document_structure",
-        std::string(
-            "Transform the currently-open document's structure using a JSON command "
-            "sequence. Supports Impress slide operations, Writer/Calc content control "
-            "updates, and arbitrary UNO commands.\n\n")
-            + DocumentToolDescriptions::TRANSFORM_PARAM_DESCRIPTION,
-        makeParamSchema(
-            {{"transform", {"string", "JSON transformation commands"}},
-             {"summary", {"string",
-                "Markdown summary of the changes for the user to review before "
-                "approving. List each slide with its title and "
-                "key content points."}}},
-            {"transform"})));
-
-    // extract_link_targets - get link targets from the open document
-    tools->add(makeAITool(
-        "extract_link_targets",
-        DocumentToolDescriptions::EXTRACT_LINK_TARGETS_DESCRIPTION,
-        makeParamSchema({}, {})));
-
-    // list_calc_functions - discover available spreadsheet functions
-    tools->add(makeAITool(
-        "list_calc_functions",
-        "List all available spreadsheet functions in the current Calc document, "
-        "grouped by category. Returns function names and signatures. "
-        "Call this when you need to verify a function exists or discover "
-        "the right function for a task. Only works for Calc/spreadsheet documents.",
-        makeParamSchema({}, {})));
-
-    // evaluate_formula - test a formula without inserting it
-    tools->add(makeAITool(
-        "evaluate_formula",
-        "Evaluate a formula without inserting it into the spreadsheet. "
-        "Returns the computed result so you can verify correctness before inserting. "
-        "Always call this before set_cell_formula to check your formula produces "
-        "the expected result. Uses US English syntax (comma separators).",
-        makeParamSchema(
-            {{"cell", {"string", "Cell address for evaluation context, e.g. 'G1'"}},
-             {"formula", {"string", "The formula to evaluate, starting with ="}}},
-            {"cell", "formula"})));
-
-    // set_cell_formula - insert formulas into cells (Calc only)
-    tools->add(makeAITool(
-        "set_cell_formula",
-        "Set formulas or values in one or more cells of the currently open spreadsheet. "
-        "Use US English formula syntax (commas as argument separators, period as decimal "
-        "separator). Always prefix formulas with =. Example: =AVERAGE(A1:A10). "
-        "Can also set plain text or numbers. Only works for Calc/spreadsheet documents.\n\n"
-        "For a single cell, provide 'cell' and 'formula' parameters.\n"
-        "For multiple cells, provide 'formulas' as a JSON array of objects, each with "
-        "'cell' and 'formula' keys. Example: [{\"cell\":\"E1\",\"formula\":\"Total\"}, "
-        "{\"cell\":\"E2\",\"formula\":\"=SUM(A2:D2)\"}]\n"
-        "Always prefer the batch 'formulas' array when setting more than one cell.",
-        makeParamSchema(
-            {{"cell", {"string", "Target cell address for a single cell, e.g. 'A1', 'B5'"}},
-             {"formula", {"string", "The formula or value for a single cell"}},
-             {"formulas", {"string", "JSON array of {cell, formula} objects for batch operations"}},
-             {"summary", {"string",
-                 "Brief human-readable description of the changes, shown to the user for approval"}}},
-            {})));
-
-    return tools;
-}
-
-bool ClientSession::handleAIChatAction(const std::string& firstLine)
-{
-    static constexpr size_t MAX_AI_PAYLOAD_SIZE = 5 * 1024 * 1024; // 5MB
-    static constexpr size_t MAX_AI_MESSAGE_LENGTH = 100 * 1024; // 100KB per message
-    static constexpr unsigned MAX_AI_MESSAGES = 50;
-
-    // Extract JSON payload after "aichat: "
-    const std::string jsonPayload = firstLine.substr(strlen("aichat: "));
-
-    if (jsonPayload.size() > MAX_AI_PAYLOAD_SIZE)
-    {
-        sendAIChatResult(false, "Request too large", "");
-        return true;
-    }
-
-    Poco::JSON::Object::Ptr requestObj = new Poco::JSON::Object();
-    if (!JsonUtil::parseJSON(jsonPayload, requestObj))
-    {
-        sendAIChatResult(false, "Invalid request format", "");
-        return true;
-    }
-
-    std::string requestId;
-    JsonUtil::findJSONValue(requestObj, "requestId", requestId);
-
-    std::string docType;
-    JsonUtil::findJSONValue(requestObj, "docType", docType);
-
-    if (docType != "text" && docType != "spreadsheet" && docType != "presentation" && docType != "drawing")
-        docType.clear();
-
-    Poco::JSON::Array::Ptr messages = requestObj->getArray("messages");
-    if (!messages || messages->size() == 0)
-    {
-        sendAIChatResult(false, "No messages provided", requestId);
-        return true;
-    }
-
-    // Build system prompt with document-type context
-    std::string systemPrompt = AI_SYSTEM_PROMPT;
-    if (!docType.empty())
-        systemPrompt += " You are currently working with a " + docType + " document.";
-
-    if (docType == "spreadsheet")
-        systemPrompt +=
-            " When referencing specific spreadsheet cells in your responses, "
-            "format them as clickable links using this pattern: [B2](cell://B2), "
-            "where the column letters come from the header row and the row number from the Row column."
-            " When the user asks for a formula, use the set_cell_formula tool to insert it directly"
-            " into the spreadsheet. Always use US English formula syntax with commas as argument"
-            " separators (e.g., =VLOOKUP(A1,B:C,2,FALSE) not =VLOOKUP(A1;B:C;2;FALSE))."
-            " Use standard Excel/Calc function names: SUM, AVERAGE, VLOOKUP, IF, COUNTIF,"
-            " SUMIF, INDEX, MATCH, etc."
-            " If you are unsure whether a function exists, call list_calc_functions to check."
-            " If the user has selected spreadsheet data, use the cell addresses visible in that data"
-            " to construct accurate cell references. If no target cell is specified by the user,"
-            " choose a sensible empty cell near the data (e.g., below the last row or to the right)."
-            " When setting multiple cells, always use a single set_cell_formula call with the"
-            " 'formulas' array parameter to batch all cells into one operation."
-            " Before inserting formulas with set_cell_formula, call evaluate_formula first"
-            " to verify the result is correct. If the result is unexpected, fix the formula"
-            " and evaluate again before inserting.";
-
-    if (docType == "presentation")
-        systemPrompt +=
-            " When creating or modifying slides, you MUST format every slide:"
-            " bold every title using EditTextObject with .uno:Bold,"
-            " apply .uno:DefaultBullet to content placeholders that list multiple items,"
-            " and choose the most appropriate layout for each slide's content from the"
-            " Available layouts list in the tool description."
-            " Do not use the same layout for every slide."
-            " Do not prefix lines with '- ' when using DefaultBullet (the bullet marker"
-            " is added automatically)."
-            " In content placeholders, put only the items to be bulleted. Do not add"
-            " sub-headings or blank lines before the bullet items."
-            " When calling transform_document_structure, include a 'summary' parameter"
-            " with a markdown preview of ONLY the slides being created or modified in"
-            " this transform call, not pre-existing slides.";
-
-    Poco::JSON::Array::Ptr sanitizedMessages = new Poco::JSON::Array();
-
-    for (unsigned i = 0; i < messages->size(); ++i)
-    {
-        auto msg = messages->getObject(i);
-        if (!msg)
-            continue;
-
-        std::string role;
-        JsonUtil::findJSONValue(msg, "role", role);
-
-        // Only allow user and assistant roles
-        if (role != "user" && role != "assistant")
-            continue;
-
-        std::string content;
-        JsonUtil::findJSONValue(msg, "content", content);
-        if (content.size() > MAX_AI_MESSAGE_LENGTH)
-        {
-            sendAIChatResult(false, "Message too long", requestId);
-            return true;
-        }
-
-        sanitizedMessages->add(msg);
-    }
-
-    // Check whether the last user message includes selected text from the document.
-    bool hasSelectedText = false;
-    for (int i = static_cast<int>(messages->size()) - 1; i >= 0; --i)
-    {
-        auto msg = messages->getObject(i);
-        if (!msg)
-            continue;
-        std::string role;
-        JsonUtil::findJSONValue(msg, "role", role);
-        if (role == "user")
-        {
-            std::string content;
-            JsonUtil::findJSONValue(msg, "content", content);
-            if (content.find("[Selected text from document:") != std::string::npos)
-                hasSelectedText = true;
-            break;
-        }
-    }
-
-    systemPrompt +=
-        " You have tools to inspect and modify the document."
-        " Use transform_document_structure to make changes.";
-
-    if (hasSelectedText)
-        systemPrompt +=
-            " The user has shared selected text from the document as context."
-            " Use that context directly to answer their question or make changes."
-            " Only call extract_document_structure if you need information about"
-            " parts of the document beyond the selection.";
-    else
-        systemPrompt +=
-            " Use extract_document_structure when you need to understand the"
-            " existing document layout before making changes."
-            " When the user asks you to create new content from scratch (like a table"
-            " or text), just generate it directly without extracting first.";
-
-    // Prepend system message: build a new array with system first, then the rest.
-    Poco::JSON::Array::Ptr finalMessages = new Poco::JSON::Array();
-    Poco::JSON::Object::Ptr systemMsg = new Poco::JSON::Object();
-    systemMsg->set("role", "system");
-    systemMsg->set("content", systemPrompt);
-    finalMessages->add(systemMsg);
-    for (unsigned i = 0; i < sanitizedMessages->size(); ++i)
-        finalMessages->add(sanitizedMessages->get(i));
-    sanitizedMessages = finalMessages;
-
-    // Trim to most recent messages if over limit (keep system prompt at index 0)
-    while (sanitizedMessages->size() > MAX_AI_MESSAGES + 1)
-        sanitizedMessages->remove(1);
-
-    // Get AI provider settings
-    const std::string apiKey = getAIProviderAPIKey();
-    const std::string model = getAIProviderModel();
-    std::string baseUrl = getAIProviderURL();
-
-    if (apiKey.empty() || model.empty() || baseUrl.empty())
-    {
-        sendAIChatResult(false, "AI settings not configured", requestId);
-        return true;
-    }
-
-    if (!baseUrl.empty() && baseUrl.back() == '/')
-        baseUrl.pop_back();
-
-    LOG_DBG("AIChatAction: request [" << requestId << "] with "
-            << sanitizedMessages->size() << " messages, model: " << model);
-
-    std::string requestUrl = std::move(baseUrl);
-    requestUrl.append("/v1/chat/completions");
-
-    // Initialize the tool loop state
-    _aiToolLoop = std::make_unique<AIToolLoopState>();
-    _aiToolLoop->requestId = requestId;
-    _aiToolLoop->messages = sanitizedMessages;
-    _aiToolLoop->model = model;
-    _aiToolLoop->requestUrl = requestUrl;
-    _aiToolLoop->apiKey = apiKey;
-
-    callLLMAPI();
-    return true;
-}
-
-void ClientSession::callLLMAPI()
-{
-    if (!_aiToolLoop)
-        return;
-
-    Poco::JSON::Object::Ptr payload = new Poco::JSON::Object();
-    payload->set("model", _aiToolLoop->model);
-    payload->set("messages", _aiToolLoop->messages);
-    payload->set("tools", buildAIToolDefinitions());
-
-    std::ostringstream payloadStream;
-    payload->stringify(payloadStream);
-    std::string payloadStr = payloadStream.str();
-
-    std::shared_ptr<http::Session> httpSession =
-        http::Session::create(_aiToolLoop->requestUrl);
-    if (!httpSession)
-    {
-        LOG_WRN("AIToolLoop: failed to create HTTP session");
-        sendAIChatResult(false, "Failed to create HTTP session", _aiToolLoop->requestId);
-        _aiToolLoop.reset();
-        return;
-    }
-
-    httpSession->setTimeout(std::chrono::seconds(getAIRequestTimeoutSeconds()));
-
-    auto clientSessionPtr = client_from_this();
-
-    http::Session::FinishedCallback finishedCallback =
-        [clientSessionPtr](const std::shared_ptr<http::Session>& session)
-    {
-        clientSessionPtr->_activeAIChatSession.reset();
-
-        if (!clientSessionPtr->_aiToolLoop)
-            return;
-
-        const std::string& requestId = clientSessionPtr->_aiToolLoop->requestId;
-        const std::shared_ptr<const http::Response> httpResponse = session->response();
-        const http::StatusCode statusCode = httpResponse->statusLine().statusCode();
-
-        if (statusCode == http::StatusCode::None)
-        {
-            clientSessionPtr->sendAIChatResult(false, "Request timeout", requestId);
-            clientSessionPtr->_aiToolLoop.reset();
-            return;
-        }
-
-        if (statusCode != http::StatusCode::OK)
-        {
-            const std::string& body = httpResponse->getBody();
-            std::cerr << "AIToolLoop: API returned "
-                << static_cast<int>(statusCode) << ' '
-                << httpResponse->statusLine().reasonPhrase()
-                << " body: " << body.substr(0, 500) << std::endl;
-            const std::string errorMessage = mapAIHttpStatusToError(
-                statusCode, httpResponse->statusLine().reasonPhrase());
-            clientSessionPtr->sendAIChatResult(false, errorMessage, requestId);
-            clientSessionPtr->_aiToolLoop.reset();
-            return;
-        }
-
-        clientSessionPtr->handleLLMResponse(httpResponse->getBody());
-    };
-
-    httpSession->setFinishedHandler(std::move(finishedCallback));
-
-    http::Session::ConnectFailCallback connectFailCallback =
-        [clientSessionPtr](const std::shared_ptr<http::Session>& /*session*/)
-    {
-        clientSessionPtr->_activeAIChatSession.reset();
-        if (clientSessionPtr->_aiToolLoop)
-        {
-            clientSessionPtr->sendAIChatResult(
-                false, "Network error - please check your connection",
-                clientSessionPtr->_aiToolLoop->requestId);
-            clientSessionPtr->_aiToolLoop.reset();
-        }
-    };
-    httpSession->setConnectFailHandler(std::move(connectFailCallback));
-
-    http::Request httpRequest(Poco::URI(_aiToolLoop->requestUrl).getPathAndQuery());
-    httpRequest.setVerb(http::Request::VERB_POST);
-    httpRequest.set("Content-Type", "application/json");
-    std::string authHeader = "Bearer ";
-    authHeader.append(_aiToolLoop->apiKey);
-    httpRequest.set("Authorization", std::move(authHeader));
-    httpRequest.setBody(std::move(payloadStr), "application/json");
-
-    LOG_DBG("AIToolLoop: sending request [" << _aiToolLoop->requestId
-            << "] round " << (6 - _aiToolLoop->toolRoundsRemaining)
-            << " to " << _aiToolLoop->requestUrl);
-
-    _activeAIChatSession = httpSession;
-    std::shared_ptr<DocumentBroker> docBroker = getDocumentBroker();
-    httpSession->asyncRequest(httpRequest, docBroker->getPoll());
-}
-
-void ClientSession::handleLLMResponse(const std::string& responseBody)
-{
-    if (!_aiToolLoop)
-        return;
-
-    const std::string& requestId = _aiToolLoop->requestId;
-
-    Poco::JSON::Object::Ptr responseObject = new Poco::JSON::Object();
-    if (!JsonUtil::parseJSON(responseBody, responseObject))
-    {
-        sendAIChatResult(false, "No response from AI", requestId);
-        _aiToolLoop.reset();
-        return;
-    }
-
-    Poco::JSON::Array::Ptr choices = responseObject->getArray("choices");
-    if (!choices || choices->size() == 0)
-    {
-        sendAIChatResult(false, "No response from AI", requestId);
-        _aiToolLoop.reset();
-        return;
-    }
-
-    Poco::JSON::Object::Ptr choice = choices->getObject(0);
-    if (!choice)
-    {
-        sendAIChatResult(false, "No response from AI", requestId);
-        _aiToolLoop.reset();
-        return;
-    }
-
-    std::string finishReason;
-    JsonUtil::findJSONValue(choice, "finish_reason", finishReason);
-
-    Poco::JSON::Object::Ptr message = choice->getObject("message");
-    if (!message)
-    {
-        sendAIChatResult(false, "No response from AI", requestId);
-        _aiToolLoop.reset();
-        return;
-    }
-
-    // Check for tool calls
-    Poco::JSON::Array::Ptr toolCalls = message->getArray("tool_calls");
-    if (toolCalls && toolCalls->size() > 0)
-    {
-        if (_aiToolLoop->toolRoundsRemaining <= 0)
-        {
-            sendAIChatResult(false, "AI used too many tool steps", requestId);
-            _aiToolLoop.reset();
-            return;
-        }
-        --_aiToolLoop->toolRoundsRemaining;
-
-        // Append the assistant message (with tool_calls) to the conversation
-        _aiToolLoop->messages->add(message);
-
-        // Capture the AI's text content so it can be shown in approval dialogs.
-        std::string assistantContent;
-        JsonUtil::findJSONValue(message, "content", assistantContent);
-        if (!assistantContent.empty())
-            _aiToolLoop->pendingSummary = assistantContent;
-
-        // Queue all tool calls for sequential processing
-        _aiToolLoop->pendingToolCalls.clear();
-        for (unsigned i = 0; i < toolCalls->size(); ++i)
-        {
-            Poco::JSON::Object::Ptr call = toolCalls->getObject(i);
-            if (!call)
-                continue;
-
-            PendingToolCall pending;
-            JsonUtil::findJSONValue(call, "id", pending.toolCallId);
-
-            Poco::JSON::Object::Ptr fn = call->getObject("function");
-            if (!fn)
-                continue;
-
-            JsonUtil::findJSONValue(fn, "name", pending.functionName);
-            JsonUtil::findJSONValue(fn, "arguments", pending.arguments);
-            _aiToolLoop->pendingToolCalls.push_back(std::move(pending));
-        }
-
-        // Start processing the first queued tool call
-        processNextPendingToolCall();
-        return;
-    }
-
-    // No tool calls - this is the final text response
-    std::string result;
-    std::string reasoning;
-    JsonUtil::findJSONValue(message, "content", result);
-    JsonUtil::findJSONValue(message, "reasoning", reasoning);
-
-    if (result.empty())
-    {
-        if (!reasoning.empty())
-        {
-            sendAIChatResult(false,
-                "This model returned only internal reasoning and no output. Try a "
-                "different model or shorter input.", requestId);
-        }
-        else if (finishReason == "length")
-        {
-            sendAIChatResult(false,
-                "The model ran out of tokens before producing output. Try a "
-                "shorter input or a model with a larger output budget.", requestId);
-        }
-        else
-        {
-            sendAIChatResult(false, "No response from AI", requestId);
-        }
-        _aiToolLoop.reset();
-        return;
-    }
-
-    sendAIChatResult(true, result, requestId);
-    _aiToolLoop.reset();
-}
-
-bool ClientSession::executeAIToolCall(const std::string& toolCallId,
-                                       const std::string& fnName,
-                                       const std::string& argsJson)
-{
-    if (!_aiToolLoop)
-        return false;
-
-    const std::string requestId = _aiToolLoop->requestId;
-
-    // generate_image - delegate to existing handler (terminates tool loop)
-    if (fnName == "generate_image")
-    {
-        Poco::JSON::Object::Ptr argsObj = new Poco::JSON::Object();
-        std::string imagePrompt;
-        if (JsonUtil::parseJSON(argsJson, argsObj))
-            JsonUtil::findJSONValue(argsObj, "prompt", imagePrompt);
-
-        if (imagePrompt.empty())
-        {
-            sendAIChatResult(false, "Image generation failed: no prompt from model", requestId);
-            _aiToolLoop.reset();
-            return true;
-        }
-
-        _aiToolLoop.reset(); // image generation is terminal
-        handleAIImageGeneration(imagePrompt, requestId);
-        return true;
-    }
-
-    std::shared_ptr<DocumentBroker> docBroker = getDocumentBroker();
-    if (!docBroker)
-    {
-        sendAIChatResult(false, "Document not available", requestId);
-        _aiToolLoop.reset();
-        return true;
-    }
-
-    // extract_document_structure - requires user approval
-    if (fnName == "extract_document_structure")
-    {
-        std::string filter;
-        Poco::JSON::Object::Ptr argsObj = new Poco::JSON::Object();
-        if (JsonUtil::parseJSON(argsJson, argsObj))
-            JsonUtil::findJSONValue(argsObj, "filter", filter);
-
-        std::string command = "extractdocumentstructure url=interactive";
-        if (!filter.empty())
-            command += " filter=" + filter;
-
-        _aiToolLoop->awaitingApproval = true;
-        _aiToolLoop->pendingToolCallId = toolCallId;
-        _aiToolLoop->pendingToolName = fnName;
-        _aiToolLoop->pendingForwardCommand = command;
-
-        sendAIToolApproval(fnName, "");
-        return true;
-    }
-
-    // extract_link_targets - read-only, send to kit
-    if (fnName == "extract_link_targets")
-    {
-        _aiToolLoop->awaitingKitResponse = true;
-        _aiToolLoop->pendingToolCallId = toolCallId;
-        _aiToolLoop->pendingToolName = fnName;
-
-        sendAIToolProgress(fnName, "Extracting link targets...");
-        forwardToChild("extractlinktargets url=interactive", docBroker);
-        return true;
-    }
-
-    // evaluate_formula - read-only, send to kit
-    if (fnName == "evaluate_formula")
-    {
-        std::string cell, formula;
-        Poco::JSON::Object::Ptr argsObj = new Poco::JSON::Object();
-        if (JsonUtil::parseJSON(argsJson, argsObj))
-        {
-            JsonUtil::findJSONValue(argsObj, "cell", cell);
-            JsonUtil::findJSONValue(argsObj, "formula", formula);
-        }
-
-        if (cell.empty() || formula.empty())
-        {
-            continueAIToolLoop(toolCallId,
-                "{\"error\":\"Missing cell or formula parameter\"}");
-            return true;
-        }
-
-        _aiToolLoop->awaitingKitResponse = true;
-        _aiToolLoop->pendingToolCallId = toolCallId;
-        _aiToolLoop->pendingToolName = fnName;
-
-        std::string encodedFormula;
-        Poco::URI::encode(formula, "", encodedFormula);
-
-        sendAIToolProgress(fnName, "Evaluating formula...");
-        forwardToChild("commandvalues command=.uno:EvaluateFormula?cell="
-            + cell + "&formula=" + encodedFormula, docBroker);
-        return true;
-    }
-
-    // list_calc_functions - read-only, send to kit
-    if (fnName == "list_calc_functions")
-    {
-        _aiToolLoop->awaitingKitResponse = true;
-        _aiToolLoop->pendingToolCallId = toolCallId;
-        _aiToolLoop->pendingToolName = fnName;
-
-        sendAIToolProgress(fnName, "Loading function catalog...");
-        forwardToChild("commandvalues command=.uno:CalcFunctionList", docBroker);
-        return true;
-    }
-
-    // set_cell_formula - requires user approval (single or batch)
-    if (fnName == "set_cell_formula")
-    {
-        std::string cell, formula, formulasJson, summary;
-        Poco::JSON::Object::Ptr argsObj = new Poco::JSON::Object();
-        if (JsonUtil::parseJSON(argsJson, argsObj))
-        {
-            JsonUtil::findJSONValue(argsObj, "cell", cell);
-            JsonUtil::findJSONValue(argsObj, "formula", formula);
-            JsonUtil::findJSONValue(argsObj, "formulas", formulasJson);
-            JsonUtil::findJSONValue(argsObj, "summary", summary);
-        }
-
-        // Build a JSON array of {cell, formula} pairs for uniform handling.
-        Poco::JSON::Array::Ptr pairs = new Poco::JSON::Array();
-
-        if (!formulasJson.empty())
-        {
-            // Batch mode: parse the formulas array
-            Poco::JSON::Array::Ptr parsed;
-            try
-            {
-                Poco::JSON::Parser parser;
-                auto result = parser.parse(formulasJson);
-                parsed = result.extract<Poco::JSON::Array::Ptr>();
-            }
-            catch (const std::exception&)
-            {
-                continueAIToolLoop(toolCallId,
-                    "{\"error\":\"Invalid JSON in formulas parameter. "
-                    "Must be an array of {cell, formula} objects.\"}");
-                return true;
-            }
-
-            for (unsigned i = 0; i < parsed->size(); ++i)
-            {
-                auto obj = parsed->getObject(i);
-                if (!obj) continue;
-                std::string c, f;
-                JsonUtil::findJSONValue(obj, "cell", c);
-                JsonUtil::findJSONValue(obj, "formula", f);
-                if (c.empty() || f.empty()) continue;
-                Poco::JSON::Object::Ptr pair = new Poco::JSON::Object();
-                pair->set("cell", c);
-                pair->set("formula", f);
-                pairs->add(pair);
-            }
-        }
-        else if (!cell.empty() && !formula.empty())
-        {
-            // Single cell mode
-            Poco::JSON::Object::Ptr pair = new Poco::JSON::Object();
-            pair->set("cell", cell);
-            pair->set("formula", formula);
-            pairs->add(pair);
-        }
-
-        if (pairs->size() == 0)
-        {
-            continueAIToolLoop(toolCallId,
-                "{\"error\":\"No valid cell/formula pairs provided. "
-                "Use 'cell'+'formula' for one cell or 'formulas' array for batch.\"}");
-            return true;
-        }
-
-        // Build summary for approval UI
-        if (summary.empty())
-        {
-            summary = "";
-            for (unsigned i = 0; i < pairs->size(); ++i)
-            {
-                auto p = pairs->getObject(i);
-                std::string c, f;
-                JsonUtil::findJSONValue(p, "cell", c);
-                JsonUtil::findJSONValue(p, "formula", f);
-                summary += "- ";
-                summary += c;
-                summary += ": `";
-                summary += f;
-                summary += "`\n";
-            }
-        }
-
-        _aiToolLoop->awaitingApproval = true;
-        _aiToolLoop->pendingToolCallId = toolCallId;
-        _aiToolLoop->pendingToolName = fnName;
-        _aiToolLoop->pendingSummary = summary;
-
-        // Store the pairs array for execution after approval.
-        std::ostringstream storedJson;
-        pairs->stringify(storedJson);
-        _aiToolLoop->pendingTransformArgs = storedJson.str();
-
-        sendAIToolApproval(fnName, "");
-        return true;
-    }
-
-    // transform_document_structure - requires user approval
-    if (fnName == "transform_document_structure")
-    {
-        std::string transform;
-        std::string summary;
-        Poco::JSON::Object::Ptr argsObj = new Poco::JSON::Object();
-        if (JsonUtil::parseJSON(argsJson, argsObj))
-        {
-            JsonUtil::findJSONValue(argsObj, "transform", transform);
-            JsonUtil::findJSONValue(argsObj, "summary", summary);
-        }
-
-        if (transform.empty())
-        {
-            continueAIToolLoop(toolCallId,
-                "{\"error\":\"No transform parameter provided\"}");
-            return true;
-        }
-
-        // findJSONValue resolves JSON escapes, so literal control characters
-        // (newlines, tabs) may appear inside string values. Escape them so
-        // the inner JSON is valid, then re-serialize through Poco for a
-        // clean string that core can parse.
-        {
-            std::string sanitized;
-            sanitized.reserve(transform.size());
-            bool inStr = false;
-            for (std::size_t i = 0; i < transform.size(); ++i)
-            {
-                char c = transform[i];
-                if (c == '"' && (i == 0 || transform[i - 1] != '\\'))
-                    inStr = !inStr;
-                if (inStr)
-                {
-                    if (c == '\n') { sanitized += "\\n"; continue; }
-                    if (c == '\r') { sanitized += "\\r"; continue; }
-                    if (c == '\t') { sanitized += "\\t"; continue; }
-                }
-                sanitized += c;
-            }
-            transform = sanitized;
-        }
-
-        Poco::JSON::Object::Ptr transformObj = new Poco::JSON::Object();
-        if (JsonUtil::parseJSON(transform, transformObj))
-        {
-            std::ostringstream oss;
-            transformObj->stringify(oss);
-            transform = oss.str();
-        }
-        else
-        {
-            continueAIToolLoop(toolCallId,
-                "{\"error\":\"Invalid JSON in transform parameter. "
-                "All slides must be in a single SlideCommands array within one "
-                "Transforms object. Use InsertMasterSlide to add slides within "
-                "the same array.\"}");
-            return true;
-        }
-
-        // Navigation-only transforms (JumpToSlide) do not modify the document
-        // and can be executed without user approval.
-        bool navigationOnly = false;
-        {
-            Poco::JSON::Object::Ptr transforms = transformObj->getObject("Transforms");
-            Poco::JSON::Array::Ptr cmds =
-                transforms ? transforms->getArray("SlideCommands") : nullptr;
-            if (cmds && cmds->size() > 0)
-            {
-                navigationOnly = true;
-                for (unsigned i = 0; i < cmds->size(); ++i)
-                {
-                    Poco::JSON::Object::Ptr cmd = cmds->getObject(i);
-                    if (!cmd ||
-                        !(cmd->has("JumpToSlide") || cmd->has("JumpToSlideByName")))
-                    {
-                        navigationOnly = false;
-                        break;
-                    }
-                }
-            }
-        }
-
-        if (navigationOnly)
-        {
-            _aiToolLoop->pendingToolCallId = toolCallId;
-            _aiToolLoop->pendingToolName = fnName;
-            _aiToolLoop->awaitingKitResponse = true;
-
-            std::string encodedTransform;
-            Poco::URI::encode(transform, "", encodedTransform);
-            forwardToChild(
-                "transformdocumentstructure url=interactive transform=" + encodedTransform,
-                docBroker);
-            return true;
-        }
-
-        _aiToolLoop->awaitingApproval = true;
-        _aiToolLoop->pendingToolCallId = toolCallId;
-        _aiToolLoop->pendingToolName = fnName;
-        _aiToolLoop->pendingTransformArgs = transform;
-        _aiToolLoop->pendingSummary = summary;
-
-        sendAIToolApproval(fnName, transform);
-        return true;
-    }
-
-    // Unknown tool - feed error back to LLM
-    continueAIToolLoop(toolCallId, "{\"error\":\"Unknown tool: " + fnName + "\"}");
-    return true;
-}
-
-void ClientSession::processNextPendingToolCall()
-{
-    if (!_aiToolLoop)
-        return;
-
-    if (_aiToolLoop->pendingToolCalls.empty())
-    {
-        // All tool calls processed, call LLM with all results
-        sendAIToolProgress(_aiToolLoop->pendingToolName, "Thinking...");
-        callLLMAPI();
-        return;
-    }
-
-    PendingToolCall next = std::move(_aiToolLoop->pendingToolCalls.front());
-    _aiToolLoop->pendingToolCalls.erase(_aiToolLoop->pendingToolCalls.begin());
-
-    LOG_DBG("AIToolLoop: tool call [" << next.functionName << "] id=" << next.toolCallId
-            << " for request [" << _aiToolLoop->requestId << ']');
-
-    executeAIToolCall(next.toolCallId, next.functionName, next.arguments);
-}
-
-void ClientSession::continueAIToolLoop(const std::string& toolCallId,
-                                        const std::string& result)
-{
-    if (!_aiToolLoop)
-        return;
-
-    // Append tool result message to the conversation
-    Poco::JSON::Object::Ptr toolResult = new Poco::JSON::Object();
-    toolResult->set("role", "tool");
-    toolResult->set("tool_call_id", toolCallId);
-    toolResult->set("content", result);
-    _aiToolLoop->messages->add(toolResult);
-
-    _aiToolLoop->awaitingKitResponse = false;
-    _aiToolLoop->awaitingApproval = false;
-
-    // Process the next queued tool call, or call LLM if all done
-    processNextPendingToolCall();
-}
-
-void ClientSession::sendAIToolProgress(const std::string& toolName,
-                                        const std::string& status)
-{
-    if (!_aiToolLoop)
-        return;
-
-    Poco::JSON::Object::Ptr progress = new Poco::JSON::Object();
-    progress->set("requestId", _aiToolLoop->requestId);
-    progress->set("toolName", toolName);
-    progress->set("status", status);
-
-    std::ostringstream oss;
-    progress->stringify(oss);
-    sendTextFrame("aichatprogress: " + oss.str());
-}
-
-void ClientSession::sendAIToolApproval(const std::string& toolName,
-                                        const std::string& transformJson)
-{
-    if (!_aiToolLoop)
-        return;
-
-    Poco::JSON::Object::Ptr approval = new Poco::JSON::Object();
-    approval->set("requestId", _aiToolLoop->requestId);
-    approval->set("toolName", toolName);
-    approval->set("transformJson", transformJson);
-    if (!_aiToolLoop->pendingSummary.empty())
-        approval->set("summary", _aiToolLoop->pendingSummary);
-
-    std::ostringstream oss;
-    approval->stringify(oss);
-    sendTextFrame("aichatapproval: " + oss.str());
-}
-
-bool ClientSession::handleAIChatApprove(const std::string& firstLine)
-{
-    const std::string jsonPayload = firstLine.substr(strlen("aichatapprove: "));
-
-    Poco::JSON::Object::Ptr obj = new Poco::JSON::Object();
-    if (!JsonUtil::parseJSON(jsonPayload, obj))
-    {
-        LOG_WRN("AIChatApprove: invalid JSON");
-        return true;
-    }
-
-    std::string action;
-    JsonUtil::findJSONValue(obj, "action", action);
-
-    if (!_aiToolLoop || !_aiToolLoop->awaitingApproval)
-    {
-        LOG_WRN("AIChatApprove: no pending approval");
-        return true;
-    }
-
-    const std::string toolCallId = _aiToolLoop->pendingToolCallId;
-
-    if (action == "approve")
-    {
-        std::shared_ptr<DocumentBroker> docBroker = getDocumentBroker();
-        if (!docBroker)
-        {
-            sendAIChatResult(false, "Document not available", _aiToolLoop->requestId);
-            _aiToolLoop.reset();
-            return true;
-        }
-
-        std::string command;
-        if (!_aiToolLoop->pendingForwardCommand.empty())
-        {
-            // Generic forwarding (extract_document_structure, etc.)
-            command = _aiToolLoop->pendingForwardCommand;
-            _aiToolLoop->pendingForwardCommand.clear();
-
-            _aiToolLoop->awaitingApproval = false;
-            _aiToolLoop->awaitingKitResponse = true;
-
-            sendAIToolProgress(_aiToolLoop->pendingToolName, "Working...");
-            forwardToChild(command, docBroker);
-        }
-        else if (_aiToolLoop->pendingToolName == "set_cell_formula")
-        {
-            _aiToolLoop->awaitingApproval = false;
-
-            // Parse the stored array of {cell, formula} pairs
-            Poco::JSON::Array::Ptr pairs;
-            try
-            {
-                Poco::JSON::Parser parser;
-                auto result = parser.parse(_aiToolLoop->pendingTransformArgs);
-                pairs = result.extract<Poco::JSON::Array::Ptr>();
-            }
-            catch (const std::exception&)
-            {
-                continueAIToolLoop(toolCallId, "{\"error\":\"Internal error parsing stored formulas\"}");
-                return true;
-            }
-
-            sendAIToolProgress("set_cell_formula", "Setting formulas...");
-
-            // Dispatch GoToCell + EnterString for each pair
-            Poco::JSON::Array resultArr;
-            for (unsigned i = 0; i < pairs->size(); ++i)
-            {
-                auto p = pairs->getObject(i);
-                if (!p) continue;
-                std::string cell, formula;
-                JsonUtil::findJSONValue(p, "cell", cell);
-                JsonUtil::findJSONValue(p, "formula", formula);
-
-                std::string escapedCell = JsonUtil::escapeJSONValue(cell);
-                std::string escapedFormula = JsonUtil::escapeJSONValue(formula);
-
-                std::string goToArgs = "{\"ToPoint\":{\"type\":\"string\",\"value\":\""
-                    + escapedCell + "\"}}";
-                forwardToChild("uno .uno:GoToCell " + goToArgs, docBroker);
-
-                std::string enterArgs = "{\"StringName\":{\"type\":\"string\",\"value\":\""
-                    + escapedFormula + "\"}}";
-                forwardToChild("uno .uno:EnterString " + enterArgs, docBroker);
-
-                Poco::JSON::Object::Ptr r = new Poco::JSON::Object();
-                r->set("cell", cell);
-                r->set("formula", formula);
-                resultArr.add(r);
-            }
-
-            // Continue tool loop with success
-            std::ostringstream resultJson;
-            Poco::JSON::Object resultObj;
-            resultObj.set("success", true);
-            resultObj.set("cells", resultArr);
-            resultObj.stringify(resultJson);
-            continueAIToolLoop(toolCallId, resultJson.str());
-        }
-        else
-        {
-            // transform_document_structure - check for GenerateImage commands
-            _aiToolLoop->awaitingApproval = false;
-            processTransformImageGenerations(docBroker);
-        }
-    }
-    else
-    {
-        // User rejected - feed rejection back to LLM with tool-specific message
-        _aiToolLoop->awaitingApproval = false;
-        std::string rejectionMsg;
-        if (_aiToolLoop->pendingToolName == "extract_document_structure")
-            rejectionMsg =
-                "{\"error\":\"User declined document inspection. "
-                "Answer their request directly without inspecting the document. "
-                "If the request is to create new content, just generate it.\"}";
-        else if (_aiToolLoop->pendingToolName == "set_cell_formula")
-            rejectionMsg =
-                "{\"error\":\"User rejected the formula insertion. "
-                "Show them the formula in a code block so they can copy it manually, "
-                "and ask if they would like a different formula.\"}";
-        else
-            rejectionMsg =
-                "{\"error\":\"User rejected the document modification. "
-                "Explain what you wanted to do and ask if they would like a different approach.\"}";
-        continueAIToolLoop(toolCallId, rejectionMsg);
-    }
-
-    return true;
-}
-
-bool ClientSession::handleAIChatCancel(const std::string& firstLine)
-{
-    const std::string cancelRequestId = firstLine.substr(strlen("aichatcancel: "));
-    LOG_DBG("AIChatCancel: cancelling request [" << cancelRequestId << ']');
-
-    if (_activeAIChatSession)
-    {
-        _activeAIChatSession->asyncShutdown();
-        _activeAIChatSession.reset();
-    }
-
-    _aiToolLoop.reset();
-
-    return true;
-}
-
-ImageGenRequest ClientSession::createImageGenRequest(const std::string& prompt)
-{
-    ImageGenRequest req;
-
-    req.apiKey = getAIImageProviderAPIKey();
-    if (req.apiKey.empty())
-        req.apiKey = getAIProviderAPIKey();
-    std::string baseUrl = getAIImageProviderURL();
-    if (baseUrl.empty())
-        baseUrl = getAIProviderURL();
-
-    if (req.apiKey.empty() || baseUrl.empty())
-    {
-        req.error = "AI image settings not configured";
-        return req;
-    }
-
-    if (!baseUrl.empty() && baseUrl.back() == '/')
-        baseUrl.pop_back();
-
-    req.requestUrl = baseUrl + "/v1/images/generations";
-
-    const std::string imageModel = getAIImageModel();
-    if (imageModel.empty())
-    {
-        req.error = "Image model not configured";
-        return req;
-    }
-
-    Poco::JSON::Object::Ptr payload = new Poco::JSON::Object();
-    payload->set("prompt", prompt);
-    std::string imageSize = getAIImageSize();
-    if (imageSize.empty() || !isValidImageSize(imageSize))
-        imageSize = "1024x1024";
-    payload->set("size", imageSize);
-    payload->set("n", 1);
-    payload->set("response_format", "b64_json");
-    payload->set("model", imageModel);
-
-    LOG_DBG("AIImageGeneration: model=" << imageModel << ", size=" << imageSize);
-
-    std::ostringstream payloadStream;
-    payload->stringify(payloadStream);
-    req.payloadStr = payloadStream.str();
-
-    req.httpSession = http::Session::create(req.requestUrl);
-    if (!req.httpSession)
-    {
-        req.error = "Failed to create HTTP session";
-        return req;
-    }
-
-    req.httpSession->setTimeout(std::chrono::seconds(getAIRequestTimeoutSeconds()));
-    return req;
-}
-
-std::pair<std::string, std::string> ClientSession::parseImageGenResponse(
-    const std::shared_ptr<const http::Response>& httpResponse)
-{
-    const http::StatusCode statusCode = httpResponse->statusLine().statusCode();
-
-    if (statusCode == http::StatusCode::None)
-        return {"", "Request timeout"};
-
-    if (statusCode != http::StatusCode::OK)
-    {
-        const std::string& body = httpResponse->getBody();
-        LOG_WRN_S("AIImageGeneration: HTTP " << static_cast<int>(statusCode) << ": " << body);
-
-        std::string errorMsg = "HTTP " + std::to_string(static_cast<int>(statusCode));
-        Poco::JSON::Object::Ptr errObj;
-        if (JsonUtil::parseJSON(body, errObj))
-        {
-            Poco::JSON::Object::Ptr errorDetail = errObj->getObject("error");
-            if (errorDetail)
-            {
-                std::string message;
-                JsonUtil::findJSONValue(errorDetail, "message", message);
-                if (!message.empty())
-                    errorMsg = message;
-            }
-        }
-        return {"", errorMsg};
-    }
-
-    const std::string& responseBody = httpResponse->getBody();
-    Poco::JSON::Object::Ptr responseObject = new Poco::JSON::Object();
-    if (!JsonUtil::parseJSON(responseBody, responseObject))
-        return {"", "Failed to parse response"};
-
-    Poco::JSON::Array::Ptr dataArray = responseObject->getArray("data");
-    if (!dataArray || dataArray->size() == 0)
-        return {"", "No image generated"};
-
-    Poco::JSON::Object::Ptr firstItem = dataArray->getObject(0);
-    if (!firstItem)
-        return {"", "No image generated"};
-
-    std::string b64Json;
-    JsonUtil::findJSONValue(firstItem, "b64_json", b64Json);
-    if (b64Json.empty())
-        return {"", "No image data in response"};
-
-    return {std::move(b64Json), ""};
-}
-
-bool ClientSession::handleAIImageGeneration(const std::string& prompt,
-                                             const std::string& requestId)
-{
-    LOG_DBG("AIImageGeneration: request [" << requestId
-            << "], prompt: " << prompt);
-
-    ImageGenRequest req = createImageGenRequest(prompt);
-    if (!req.error.empty())
-    {
-        sendAIChatResult(false, req.error, requestId);
-        return true;
-    }
-
-    // Send image result via aichatresult with imageData field
-    auto clientSessionPtr = client_from_this();
-    auto sendImageResult = [clientSession = clientSessionPtr, requestId](
-                               bool success, const std::string& imageData,
-                               const std::string& error)
-    {
-        Poco::JSON::Object::Ptr result = new Poco::JSON::Object();
-        result->set("success", success);
-        if (success)
-            result->set("imageData", imageData);
-        else
-            result->set("error", error);
-        result->set("requestId", requestId);
-
-        std::ostringstream oss;
-        result->stringify(oss);
-        clientSession->sendTextFrame("aichatresult: " + oss.str());
-    };
-
-    http::Session::FinishedCallback finishedCallback =
-        [clientSessionPtr, sendImageResult](const std::shared_ptr<http::Session>& session)
-    {
-        clientSessionPtr->_activeAIChatSession.reset();
-
-        auto [b64Json, error] = parseImageGenResponse(session->response());
-        if (!error.empty())
-        {
-            sendImageResult(false, "", error);
-            return;
-        }
-
-        sendImageResult(true, b64Json, "");
-    };
-
-    req.httpSession->setFinishedHandler(std::move(finishedCallback));
-
-    http::Session::ConnectFailCallback connectFailCallback =
-        [clientSessionPtr = std::move(clientSessionPtr),
-         sendImageResult = std::move(sendImageResult)](
-            const std::shared_ptr<http::Session>& /*session*/)
-    {
-        clientSessionPtr->_activeAIChatSession.reset();
-        sendImageResult(false, "", "Network error - please check your connection");
-    };
-    req.httpSession->setConnectFailHandler(std::move(connectFailCallback));
-
-    http::Request httpRequest(Poco::URI(req.requestUrl).getPathAndQuery());
-    httpRequest.setVerb(http::Request::VERB_POST);
-    httpRequest.set("Content-Type", "application/json");
-    httpRequest.set("Authorization", "Bearer " + req.apiKey);
-    httpRequest.setBody(req.payloadStr, "application/json");
-
-    LOG_DBG("AIImageGeneration: sending request [" << requestId << "] to "
-            << req.requestUrl);
-
-    _activeAIChatSession = req.httpSession;
-    std::shared_ptr<DocumentBroker> docBroker = getDocumentBroker();
-    req.httpSession->asyncRequest(httpRequest, docBroker->getPoll());
-    return true;
-}
-
-void ClientSession::processTransformImageGenerations(
-    const std::shared_ptr<DocumentBroker>& docBroker)
-{
-    if (!_aiToolLoop)
-        return;
-
-    const std::string& transform = _aiToolLoop->pendingTransformArgs;
-
-    // Parse transform to find GenerateImage.N commands in SlideCommands
-    Poco::JSON::Object::Ptr transformObj = new Poco::JSON::Object();
-    if (!JsonUtil::parseJSON(transform, transformObj))
-    {
-        // Should not happen - already validated earlier
-        _aiToolLoop->awaitingKitResponse = true;
-        std::string encodedTransform;
-        Poco::URI::encode(transform, "", encodedTransform);
-        sendAIToolProgress(_aiToolLoop->pendingToolName, "Working...");
-        forwardToChild(
-            "transformdocumentstructure url=interactive transform=" + encodedTransform,
-            docBroker);
-        return;
-    }
-
-    Poco::JSON::Object::Ptr transforms = transformObj->getObject("Transforms");
-    Poco::JSON::Array::Ptr cmds =
-        transforms ? transforms->getArray("SlideCommands") : nullptr;
-
-    _aiToolLoop->pendingImageGens.clear();
-    _aiToolLoop->nextImageGenIndex = 0;
-    _aiToolLoop->generatingImages = false;
-    _aiToolLoop->outstandingImageTransforms = 0;
-    _aiToolLoop->mainTransformResult.clear();
-    _aiToolLoop->failedImagePrompts.clear();
-
-    // Track current slide index as we scan commands
-    int currentSlide = 0;
-    int pageCount = 1;
-
-    if (cmds)
-    {
-        for (unsigned i = 0; i < cmds->size(); ++i)
-        {
-            Poco::JSON::Object::Ptr cmd = cmds->getObject(i);
-            if (!cmd)
-                continue;
-
-            // Track slide navigation to determine which slide each
-            // GenerateImage targets
-            if (cmd->has("JumpToSlide"))
-            {
-                std::string val = cmd->getValue<std::string>("JumpToSlide");
-                if (val == "last")
-                    currentSlide = pageCount - 1;
-                else
-                {
-                    try
-                    {
-                        currentSlide = std::stoi(val);
-                    }
-                    catch (const std::exception&)
-                    {
-                        LOG_WRN("TransformImageGen: invalid JumpToSlide value: " << val);
-                    }
-                }
-            }
-            else if (cmd->has("InsertMasterSlide")
-                     || cmd->has("InsertMasterSlideByName"))
-            {
-                currentSlide++;
-                pageCount++;
-            }
-            else if (cmd->has("DeleteSlide"))
-            {
-                if (pageCount > 1)
-                    pageCount--;
-            }
-
-            static const std::string kGenerateImagePrefix = "GenerateImage.";
-            for (const auto& key : cmd->getNames())
-            {
-                if (key.substr(0, kGenerateImagePrefix.size()) == kGenerateImagePrefix)
-                {
-                    int objId;
-                    try
-                    {
-                        objId = std::stoi(key.substr(kGenerateImagePrefix.size()));
-                    }
-                    catch (const std::exception&)
-                    {
-                        LOG_WRN("TransformImageGen: invalid GenerateImage key: " << key);
-                        continue;
-                    }
-
-                    std::string prompt = cmd->getValue<std::string>(key);
-                    _aiToolLoop->pendingImageGens.push_back(
-                        {currentSlide, objId, prompt, std::string()});
-
-                    // Replace GenerateImage.N with InsertImage.N pointing
-                    // to the loading placeholder
-                    cmd->remove(key);
-                    cmd->set("InsertImage." + std::to_string(objId),
-                        "file://" + std::string(JAILED_DOCUMENT_ROOT)
-                        + "insertfile/ai-loading-placeholder.png");
-                }
-            }
-        }
-    }
-
-    if (_aiToolLoop->pendingImageGens.size() > MAX_AI_IMAGE_GENERATIONS)
-    {
-        LOG_WRN("TransformImageGen: capping image generations from "
-                << _aiToolLoop->pendingImageGens.size()
-                << " to " << MAX_AI_IMAGE_GENERATIONS);
-        _aiToolLoop->pendingImageGens.resize(MAX_AI_IMAGE_GENERATIONS);
-    }
-
-    // Copy the placeholder image to the jail insertfile directory
-    if (!_aiToolLoop->pendingImageGens.empty())
-    {
-        const std::string jailId = docBroker->getJailId();
-        const std::string dirPath = FileUtil::buildLocalPathToJail(
-            COOLWSD::EnableMountNamespaces, COOLWSD::ChildRoot + jailId,
-            std::string(JAILED_DOCUMENT_ROOT) + "insertfile");
-        Poco::File(dirPath).createDirectories();
-
-        const std::string srcPath = COOLWSD::FileServerRoot
-            + "/browser/dist/images/ai-loading-placeholder.png";
-        const std::string dstPath = dirPath + "/ai-loading-placeholder.png";
-
-        try
-        {
-            Poco::File src(srcPath);
-            if (src.exists())
-                src.copyTo(dstPath);
-            else
-                LOG_WRN("TransformImageGen: placeholder not found at " << srcPath);
-        }
-        catch (const std::exception& ex)
-        {
-            LOG_WRN("TransformImageGen: failed to copy placeholder: " << ex.what());
-        }
-    }
-
-    // Serialize the modified transform (with placeholders or unchanged)
-    std::ostringstream oss;
-    transformObj->stringify(oss);
-    std::string modifiedTransform = oss.str();
-
-    // Forward the transform immediately (with loading placeholders if any)
-    _aiToolLoop->awaitingKitResponse = true;
-    std::string encodedTransform;
-    Poco::URI::encode(modifiedTransform, "", encodedTransform);
-    sendAIToolProgress(_aiToolLoop->pendingToolName, "Working...");
-    forwardToChild(
-        "transformdocumentstructure url=interactive transform=" + encodedTransform,
-        docBroker);
-
-    // Mark that we need to generate images after the kit responds
-    if (!_aiToolLoop->pendingImageGens.empty())
-        _aiToolLoop->generatingImages = true;
-}
-
-std::string ClientSession::appendImageGenFailures(const std::string& result) const
-{
-    if (!_aiToolLoop || _aiToolLoop->failedImagePrompts.empty())
-        return result;
-
-    std::string augmented = result;
-    augmented += "\n\nNote: ";
-    augmented += std::to_string(_aiToolLoop->failedImagePrompts.size());
-    augmented += " image(s) failed to generate and still show a loading placeholder."
-                 " Failed prompts: ";
-    for (std::size_t i = 0; i < _aiToolLoop->failedImagePrompts.size(); ++i)
-    {
-        if (i > 0)
-            augmented += ", ";
-        augmented += "\"";
-        augmented += _aiToolLoop->failedImagePrompts[i];
-        augmented += "\"";
-    }
-    return augmented;
-}
-
-void ClientSession::generateNextTransformImage(std::shared_ptr<DocumentBroker> docBroker)
-{
-    if (!_aiToolLoop)
-        return;
-
-    for (std::size_t& idx = _aiToolLoop->nextImageGenIndex;
-         idx < _aiToolLoop->pendingImageGens.size(); ++idx)
-    {
-        const PendingImageGen& gen = _aiToolLoop->pendingImageGens[idx];
-        const std::size_t total = _aiToolLoop->pendingImageGens.size();
-
-        sendAIToolProgress(_aiToolLoop->pendingToolName, "Generating image " +
-                                                             std::to_string(idx + 1) + " of " +
-                                                             std::to_string(total) + "...");
-
-        ImageGenRequest req = createImageGenRequest(gen.prompt);
-        if (!req.error.empty())
-        {
-            LOG_WRN("TransformImageGen: " << req.error);
-            _aiToolLoop->failedImagePrompts.push_back(gen.prompt);
-            continue;
-        }
-
-        auto clientSessionPtr = client_from_this();
-
-        auto onImageFail = [clientSessionPtr, docBroker, idx]()
-        {
-            clientSessionPtr->_aiToolLoop->failedImagePrompts.push_back(
-                clientSessionPtr->_aiToolLoop->pendingImageGens[idx].prompt);
-            clientSessionPtr->_aiToolLoop->nextImageGenIndex++;
-            clientSessionPtr->generateNextTransformImage(docBroker);
-        };
-
-        http::Session::FinishedCallback finishedCallback =
-            [clientSessionPtr, docBroker, idx, onImageFail](
-                const std::shared_ptr<http::Session>& session)
-        {
-            clientSessionPtr->_activeAIChatSession.reset();
-
-            if (!clientSessionPtr->_aiToolLoop)
-                return;
-
-            auto [b64Json, error] = parseImageGenResponse(session->response());
-            if (!error.empty())
-            {
-                LOG_WRN_S("TransformImageGen: " << error);
-                onImageFail();
-                return;
-            }
-
-            // Decode base64 and write to jail insertfile directory
-            std::string binaryData;
-            macaron::Base64::Decode(b64Json, binaryData);
-
-            const std::string jailId = docBroker->getJailId();
-            const std::string dirPath = FileUtil::buildLocalPathToJail(
-                COOLWSD::EnableMountNamespaces, COOLWSD::ChildRoot + jailId,
-                std::string(JAILED_DOCUMENT_ROOT) + "insertfile");
-
-            Poco::File(dirPath).createDirectories();
-
-            const std::string fileName =
-                "ai_" + Util::rng::getHexString(8) + "_" + std::to_string(idx) + ".png";
-            std::string filePath = dirPath;
-            filePath += '/';
-            filePath += fileName;
-
-            std::ofstream fileStream(filePath, std::ios::out | std::ios::binary);
-            fileStream.write(binaryData.data(), binaryData.size());
-            fileStream.close();
-
-            if (!fileStream.good())
-            {
-                LOG_WRN_S("TransformImageGen: failed to write image to " << filePath);
-                FileUtil::removeFile(filePath);
-                onImageFail();
-                return;
-            }
-
-            // Build jail-internal file:// URL for core to load
-            const std::string jailFileUrl =
-                "file://" + std::string(JAILED_DOCUMENT_ROOT) + "insertfile/" + fileName;
-            clientSessionPtr->_aiToolLoop->pendingImageGens[idx].filePath = jailFileUrl;
-
-            LOG_DBG_S("TransformImageGen: wrote image " << (idx + 1) << " to " << filePath);
-
-            // Send a transform to replace the loading placeholder with the real
-            // image on the correct slide.
-            const auto& imgGen = clientSessionPtr->_aiToolLoop->pendingImageGens[idx];
-            std::string miniTransform = "{\"Transforms\":{\"SlideCommands\":["
-                                        "{\"InsertImageAt." +
-                                        std::to_string(imgGen.slideIndex) + "." +
-                                        std::to_string(imgGen.objId) + "\":\"" + jailFileUrl +
-                                        "\"}]}}";
-
-            std::string encodedMini;
-            Poco::URI::encode(miniTransform, "", encodedMini);
-            clientSessionPtr->_aiToolLoop->outstandingImageTransforms++;
-            clientSessionPtr->forwardToChild(
-                "transformdocumentstructure url=interactive transform=" + encodedMini, docBroker);
-
-            clientSessionPtr->_aiToolLoop->nextImageGenIndex++;
-            clientSessionPtr->generateNextTransformImage(docBroker);
-        };
-
-        req.httpSession->setFinishedHandler(std::move(finishedCallback));
-
-        http::Session::ConnectFailCallback connectFailCallback =
-            [clientSessionPtr, onImageFail](const std::shared_ptr<http::Session>& /*session*/)
-        {
-            clientSessionPtr->_activeAIChatSession.reset();
-
-            if (!clientSessionPtr->_aiToolLoop)
-                return;
-
-            LOG_WRN_S("TransformImageGen: connection failed");
-            onImageFail();
-        };
-        req.httpSession->setConnectFailHandler(std::move(connectFailCallback));
-
-        http::Request httpRequest(Poco::URI(req.requestUrl).getPathAndQuery());
-        httpRequest.setVerb(http::Request::VERB_POST);
-        httpRequest.set("Content-Type", "application/json");
-        httpRequest.set("Authorization", "Bearer " + req.apiKey);
-        httpRequest.setBody(req.payloadStr, "application/json");
-
-        LOG_DBG("TransformImageGen: generating image " << (idx + 1) << " of " << total
-                                                       << ", prompt: " << gen.prompt);
-
-        _activeAIChatSession = req.httpSession;
-        req.httpSession->asyncRequest(httpRequest, docBroker->getPoll());
-        return; // async request launched, callbacks will call back into this function
-    }
-
-    // All entries processed or skipped - check completion
-    if (_aiToolLoop->outstandingImageTransforms > 0)
-        return; // responses still pending, they will finish up
-    _aiToolLoop->generatingImages = false;
-    continueAIToolLoop(_aiToolLoop->pendingToolCallId,
-        appendImageGenFailures(_aiToolLoop->mainTransformResult));
-}
 
 bool ClientSession::handleSignatureAction(const StringVector& tokens)
 {
@@ -2999,15 +1336,15 @@ bool ClientSession::_handleInput(const char *buffer, int length)
 #if !MOBILEAPP
     else if (tokens.equals(0, "aichat:"))
     {
-        return handleAIChatAction(firstLine);
+        return _aiChat->handleAction(firstLine);
     }
     else if (tokens.equals(0, "aichatcancel:"))
     {
-        return handleAIChatCancel(firstLine);
+        return _aiChat->handleCancel(firstLine);
     }
     else if (tokens.equals(0, "aichatapprove:"))
     {
-        return handleAIChatApprove(firstLine);
+        return _aiChat->handleApprove(firstLine);
     }
 #endif
     else if (tokens.equals(0, "resetaccesstoken"))
@@ -4193,15 +2530,8 @@ bool ClientSession::handleKitToClientMessage(const std::shared_ptr<Message>& pay
                 return false;
             }
 #if !MOBILEAPP
-            else if (_aiToolLoop && _aiToolLoop->awaitingKitResponse &&
-                     (errorCommand == "extractdocumentstructure" ||
-                      errorCommand == "extractlinktargets" ||
-                      errorCommand == "transformdocumentstructure"))
+            else if (_aiChat->tryConsumeKitError(errorCommand, errorKind))
             {
-                LOG_WRN("AIToolLoop: kit error for " << errorCommand << ": " << errorKind);
-                _aiToolLoop->awaitingKitResponse = false;
-                continueAIToolLoop(_aiToolLoop->pendingToolCallId,
-                    "{\"error\":\"" + errorCommand + " failed: " + errorKind + "\"}");
                 return true;
             }
             else
@@ -4624,15 +2954,8 @@ ClientSession::handleOpenDocKitToClientMessage(const std::shared_ptr<Message>& p
     else if (tokens.equals(0, "commandvalues:"))
     {
 #if !MOBILEAPP
-        // Intercept commandvalues responses for the AI tool loop
-        if (_aiToolLoop && _aiToolLoop->awaitingKitResponse
-            && (_aiToolLoop->pendingToolName == "list_calc_functions"
-                || _aiToolLoop->pendingToolName == "evaluate_formula"))
-        {
-            _aiToolLoop->awaitingKitResponse = false;
-            continueAIToolLoop(_aiToolLoop->pendingToolCallId, payload->jsonString());
+        if (_aiChat->tryConsumeCommandValues(payload))
             return true;
-        }
 #endif
 
         const std::string stringJSON = payload->jsonString();
@@ -4764,12 +3087,8 @@ ClientSession::handleOpenDocKitToClientMessage(const std::shared_ptr<Message>& p
     else if (tokens.equals(0, "extractedlinktargets:"))
     {
 #if !MOBILEAPP
-        if (_aiToolLoop && _aiToolLoop->awaitingKitResponse)
-        {
-            _aiToolLoop->awaitingKitResponse = false;
-            continueAIToolLoop(_aiToolLoop->pendingToolCallId, payload->jsonString());
+        if (_aiChat->tryConsumeExtractedLinkTargets(payload))
             return true;
-        }
 #endif
 
         LOG_TRC("Sending extracted link targets response.");
@@ -4792,12 +3111,8 @@ ClientSession::handleOpenDocKitToClientMessage(const std::shared_ptr<Message>& p
     else if (tokens.equals(0, "extracteddocumentstructure:"))
     {
 #if !MOBILEAPP
-        if (_aiToolLoop && _aiToolLoop->awaitingKitResponse)
-        {
-            _aiToolLoop->awaitingKitResponse = false;
-            continueAIToolLoop(_aiToolLoop->pendingToolCallId, payload->jsonString());
+        if (_aiChat->tryConsumeExtractedDocumentStructure(payload))
             return true;
-        }
 #endif
 
         LOG_TRC("Sending extracted document structure response.");
@@ -4820,65 +3135,8 @@ ClientSession::handleOpenDocKitToClientMessage(const std::shared_ptr<Message>& p
     else if (tokens.equals(0, "transformeddocumentstructure:"))
     {
 #if !MOBILEAPP
-        if (_aiToolLoop && _aiToolLoop->awaitingKitResponse)
-        {
-            _aiToolLoop->awaitingKitResponse = false;
-
-            // If we have pending image generations, start generating them
-            // instead of continuing the AI tool loop. The main transform
-            // (with loading placeholders) has been applied - now we
-            // progressively replace placeholders with real images.
-            if (_aiToolLoop->generatingImages)
-            {
-                _aiToolLoop->mainTransformResult = payload->jsonString();
-                std::shared_ptr<DocumentBroker> broker = getDocumentBroker();
-                if (broker)
-                    generateNextTransformImage(broker);
-                else
-                {
-                    _aiToolLoop->generatingImages = false;
-                    continueAIToolLoop(_aiToolLoop->pendingToolCallId,
-                        payload->jsonString());
-                }
-                return true;
-            }
-
-            continueAIToolLoop(_aiToolLoop->pendingToolCallId, payload->jsonString());
+        if (_aiChat->tryConsumeTransformedDocumentStructure(payload))
             return true;
-        }
-#endif
-
-#if !MOBILEAPP
-        // During progressive image generation, image insertion transforms
-        // produce responses that reach here. Track their success and, when
-        // all responses are in, continue the AI tool loop.
-        if (_aiToolLoop && _aiToolLoop->generatingImages)
-        {
-            if (_aiToolLoop->outstandingImageTransforms > 0)
-            {
-                _aiToolLoop->outstandingImageTransforms--;
-
-                const std::string jsonResult = payload->jsonString();
-                Poco::JSON::Object::Ptr resultObj;
-                if (JsonUtil::parseJSON(jsonResult, resultObj))
-                {
-                    bool success = false;
-                    JsonUtil::findJSONValue(resultObj, "success", success);
-                    if (!success)
-                        LOG_WRN("Image insertion transform failed: " << jsonResult);
-                }
-            }
-
-            if (_aiToolLoop->outstandingImageTransforms <= 0
-                && _aiToolLoop->nextImageGenIndex
-                       >= _aiToolLoop->pendingImageGens.size())
-            {
-                _aiToolLoop->generatingImages = false;
-                continueAIToolLoop(_aiToolLoop->pendingToolCallId,
-                    appendImageGenFailures(_aiToolLoop->mainTransformResult));
-            }
-            return true;
-        }
 #endif
 
         LOG_TRC("Sending transformed document structure response.");
